pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monitor_mcp"
version = "0.1.0"
description = "An MCP server for controlled screen monitoring and frame retrieval."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=0.1.0",
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "streamlit>=1.24.0",
    "pyautogui>=0.9.54",
    "google-genai>=0.1.0",
]

[project.scripts]
monitor-mcp = "monitor_mcp.server:main"
monitor-mcp-ui = "monitor_mcp.ui:main"

[tool.hatch.build.targets.wheel]
packages = ["src/monitor_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
